Changed the rules file to suite for unit test
This change is required because in unit tests for recording
rules we need to provide numeric inputs value and with
existing recording rule we need boolean. With this change
the recording rule still returns numeric value and we can
easily add unit tests for them. Also changed the alerting
rule to check the value and then raise the alert, rather than
just calling the recording rule.

ceph-mixins-bug-id: ceph/ceph-mixins/issues/55
Signed-off-by: Shubhendu <shtripat@redhat.com>